Support kibana version 6.3 >
Fixes problem with changes to search api

Resolves #11
