+++ .travis.yml
@@ -8,6 +8,7 @@ env:
   - ELK_VERSION=5.5.3   KIBANA_TYPE=KibanaTypeVanilla   ELASTIC_PACK=
   - ELK_VERSION=5.5.3   KIBANA_TYPE=KibanaTypeLogzio    KIBANA_URI=https://app-eu.logz.io    ELASTIC_SEARCH_PATH=/kibana/elasticsearch/logzioCustomerKibanaIndex  LOGZ_IO_ACCOUNT_ID_1=16533  LOGZ_IO_ACCOUNT_ID_2=14942
   - ELK_VERSION=6.2.1   KIBANA_TYPE=KibanaTypeVanilla
+  - ELK_VERSION=6.4.1   KIBANA_TYPE=KibanaTypeVanilla
   global:
   - secure: q5fWz4h3MJvt2eIwnAoRkNWK0RrgEcJApe/H/Ejf55hJSijX/DhHxOwjbo/TDaaCVoMkb9HSzKEhF3tDdduuRulGibvmAapyumHUa5fJmwlHwLXgrP7YiJ+4f/AcupoGRRiKOdUNwPVYo48ewaj44y/0Wu8L6Lzi09LaSs6dRAsN8IfVwhtjCxduPsXDsMNvzNv6VlW4XZkICRpxq1e4+8RK/DiWk3J5e4v3g80fzNoXx3g1J0zjcXu5RR0qlkY16f7BHilIPhaBltJS3vu0g+jKEWf7dLhjTHM6Zs2qRg9eywNjDEDVhSQkLdJXG1FNQRpsAFpIkjilC5PU3nX5Wft39QjUXQxjFtSkiwYt5HS9A56IczstNzy7hj3SGdvXM9w3ugkRW5FPwQfZ2x5gbbMikIswM+3VbZRi4cf8c3P23LBYxgN9gCHWFHyf7Ta/L4AM/Yqq0aHWbn4OgKa+V+uIhrW6I3DxlOJXAjQU3/CGrmidhAu7kfFikV+G0cXbtnyPuf5n12wcx21b5QJeOgNroLHS7cfYQEWn12wvyJImq0r6RGJxjIJBqxc9mSOu19aX7UaKlM1qEyen/UflcheKfUtzTw7knVLcYWVVX5Y07k3P/do5TRhofUSJ+DBexCkrY+3c+N78/V82JPBxpGaRdGJ1VsW8KxlO5kdXpBg=
   - secure: vDdr/9AK47R12x/a3D3f1F/rmPUlkNcLD/TUANU+b0AmKB+JV21BMDvJxeDxYYZUbjJ7pRQzjNaatdDn3DOtd7h7y6Qt7/6QOkO+ZWtpmWROFjYz76cK3YutrSQrfhPHC8CJIQO/HxsQ2BhFRc+c6a6KTdc6KzzQdAHNNN0jpyXvMwTSiAEzpGTUqVZFCW3noG621weekGmtMdfQTx3q6UxUV1RV62GzGLUD8jJqzgSH4+lEXzDR5dIqRp3AsLlkf62rC2DRiYW724Nj9jz87pkBtXhXJICKbbnHt3shLXObkj/LC+Vz+rfBT88KEz4xLOvTtkUGAvYt1FpPvwvvf9EilyB3kHOSMpaIgIohHXMSV3a8/4S09H1gdIxSdS1LjqqDswfDYZBnMN2yRiEEhnD0AUpNjSYLhA8D6rdB5/Goi/iqglNg0KhHWaogoXZ8/i0K3m4fmsSKUbhN1WixRwNKXrLmR2pFb4Ak9zxzDivRtxekRg82I1XAsZnRvjSr7EX618nhWs/bVkFJHC6dBRQK+pfTCiyVHY82av4A/OxFPBFJ5TJv+1B6vtbnQe+IGMJ1MrSn4SODNd90jwWvMib5yZqAwCKPH/rU6LOXtb2nP7RzvwxoUymsGW8V3DBBXyUQi2KJ6mi+CurZJQm8viKa2PezUaNym/QJcAPlby4=
